[platform] IDEA-275323 Refactoring: auto scroll methods/key renames, split isAutoScrollFromSource()
The methods and the key renamed consistently to ensure auto scroll from source
is mentioned, to avoid confusion with other kinds of auto scrolling behavior.

The blockAutoScrollFromSource() method is made private.

The isAutoScrollFromSource() method logic is split in two methods: one
related to the actual state of the option (enabled/disabled) and one related
to this temporary blocking of auto scrolling behavior. We'll need this to
fix Select Open File appearing briefly while auto scroll from source is blocked,
which along with already committed fixes, should fix IDEA-302838.

GitOrigin-RevId: d2e1b4c98cc7130c153be17c038919bbe0f927d7