ML in SE: Remove unnecessary sync for getting ml features and weight (IDEA-304690)
getElementFeatures and getMlWeight are stateless methods, which means
that synchronisation of them is unnecessary and parallel execution
is not going to be a problem.

The only reason why the synchronisation remained was due to previous
logic, where information was cached.

(cherry picked from commit f4a29f0e038710620cf06732a747079a33197a61)

IJ-CR-97802

GitOrigin-RevId: f63586ba516093d40461a53a0cad34bc53d6a7f1