[platform] IDEA-305285 Enable customization of New Element popups
Customization didn't work because the group that is customized is
the Project View popup group, but Alt+Ins only uses the New Element
group, which isn't a customization root, so to speak.

Fix by getting the customized Project View popup group and searching
for the New Element group there. This looks more like a workaround
than a proper fix, but at least it works. Definitely needs improvement, though.

GitOrigin-RevId: 84956ff1f68148e176446a57436e27dd83f9c9cf