settings-sync: don't go throw the common migration procedure when migrating from the Settings Repository
It appears that "migrate settings AND enable settings sync" is significantly different from "migrate settings and DON'T enable settings sync", and trying to unify them produces a lot of hacks in the initializeSync() procedure. Moreover, the partial initialization of the SettingsSyncBridge implemented previously leads to IDEA-305271.

In this commit we remove the hacks from initializeSync() and perform migration as a separate procedure. Going directly through the IdeMediator, the migration copies files to the APP_CONFIG and reloads affected components, but doesn't create a settingsSync repository.

The only downside of this approach is that it overwrites the existing settings in APP_CONFIG, and if something goes wrong (e.g. if we incorrectly detect that the migration is needed at all) then we'll erase actual configs. We mitigate this risk by creating a backup of current configs before migrating.

(cherry picked from commit 5a14ad1e9c789ad3d73c0a9716148db95bafe52a)

GitOrigin-RevId: b6e567d2ce987cb1c83324f5ae237c05f5febdbc