Fix race in DumbServiceImpl: tasks may hang in the "dumb queue" until the following `queueTask` invocation
If a task was submitted to `DumbServiceImpl.queueTask` after DumbServiceGuiTaskQueue has finished its polling in the while loop, but before `myState.set(State.WAITING_FOR_FINISH)` is invoked in the callback, resulting state will be WAITING_FOR_FINISH (followed by SMART), and task will remain in the queue until the following `queueTask` invocation.

Desired (new) behavior: resulting state is SCHEDULED_TASKS, and task will be scheduled for execution in a background thread.

Series of commits with the goal to make DumbServiceGuiTaskQueue a reusable component

GitOrigin-RevId: b5e07c6f1f746de87d6f2af844f194217a3525dc