[platform] IDEA-302128 Fix broken shortcuts for AnActionButton
Moving from AnActionButton.getShortcut() to AnAction.getShortcutSet()
caused CommonActionsPanel to ignore overrides of the getShortcut()
method, which, first of all, broke all standard (MyActionButton)
shortcuts like Alt+Insert because it's defined in the override.

For now, just use getShortcut() if the action is an instance of AnActionButton.
We can get rid of this code later, when we either get rid of AnActionButton
altogether or think of a better way to handle this.

(cherry picked from commit 59a7b911501ff8e56376adf76df2b2cfc711d2c9)

IJ-CR-98191

GitOrigin-RevId: 2aa5fcc6118be3456f40d0f3df4b7ab639512917