IDEA-306358 EA-719755 fix the race in the decoder
The only way I see when assertion may fail - effectiveLexer changed between obtaining element type and text. Meaning -  other thread advanced and/or appended a text.
This lock should make chunk processing atomic and prevent lexemes mixing on concurrent access.
This solution was not intended to be used concurrently, but probably this lock will improve things.
The other way is to make a thread control for each underlying lexer and fail immediatelly on concurrent access.

IJ-MR-98255

GitOrigin-RevId: 323efadb9dc502143e6407bcda603e8ad739717b