[platform] introduce ProjectFileIndex::isInGeneratedSources and use it instead of getSourceFolder
'getSourceFolder' delegates to non-incremental implementation of DirectoryIndex (IDEA-276394), now it's deprecated. Code which use it to determine whether file is located under generated source root, or to check type of the containing source root, migrated to use other methods (which used incremental WorkspaceFileIndex). ProjectRootsUtil.getModuleSourceRoot is left for cases where SourceFolder is really needed, and it's implementation is changed to slightly less efficient.

GitOrigin-RevId: 492a1a94ff7298f774036bb514227b1ee136ee21