[newUI] IDEA-293980 Create a setting for independent tool window sizes
For the old UI, it's still implemented as a registry key, in case someone
needs it. For the new UI, the registry key is removed and replaced with
the new setting according to UX-2272.

Implementation note: instead of using visible(ExperimentalUI.isNewUI()),
an if-else was used, because otherwise invisible check boxes would still
take up space, which looked weird (like there's something missing).

GitOrigin-RevId: d7c562f965a843e835bdbe4e7b22880de1217461