[platform] IDEA-305758 Do not hide an empty action group if it can be performed
A very edge case of expanding an action group: it's considered effectively empty
if one of the following is true:
- it has no enabled items, and we were asked to hide disabled ones;
- it has no visible items, and we were asked to hide empty groups.

However, when the group itself is performable, then this logic is flawed:
it may not have effectively visible children, but the group itself should then
be visible as a single action. This is sometimes used for groups that automatically
collapse into a single action when there is only one child: instead of showing
a submenu of one item, the group itself becomes the item and hides its children.
An example is GHOpenInBrowserActionGroup.

Fix by expanding the group to itself (instead of the empty list) when it's effectively empty,
but performable.

GitOrigin-RevId: 08e90decef4ab61bc1ae759c44aba7e3ff67b229