[platform UI] IDEA-304888 Fall back to last active editor for Select In action
When focus is in the project view, there is technically no active editor, so
the Select In action would fall back to PSI Select In context, which eventually
caused the file selected in the Project View to be used as the Select In context.
This broke the ability to select the file opened in the selected, but not focused
editor tab in the project view which is what IDEA-304888 is about.

Fix by falling back to LAST_ACTIVE_FILE_EDITOR in case there is no active
editor attached to the event data context. This has a side effect of using the
selected editor tab for ALL Select In actions, not just Select In Project View.
This could in theory break some workflow when the user selects a file in
the Project View somehow without opening it in the editor and then doing
something like Select In Structure. However, implementing this for just
Select In Project View would lead to inconsistent behavior that may or may
not be desirable, which is hard to judge at this point.

GitOrigin-RevId: e067eff95d54cbde854cdd9890c5f8a173a0ac7b