[platform UI] IDEA-304888 Use inactive selected editor for Select In only in Project View
The last fix broke a lot of Alt+F1 actions in tool windows unrelated to Project View,
such as Git (select a branch in the log).

To fix this, only fall back to the selected but not focused editor if the context component
is in the Project View. To avoid an ugly `instanceof ProjectViewTree` in SelectInContextImpl,
we reverse this dependency a bit: we declare a nested interface to retrieve what the context
component considers the active file editor, and then we put it into a client property to
avoid implementing it directly in the context component, which would be possible in
this case but generally not desirable as the component could be of a more or less standard
type.

This way components now can specify what they consider the active editor. It's only
implemented for ProjectViewTree now, other components preserve their original behavior.

GitOrigin-RevId: 27ebb6bb096401dd32c6721e065bafb7acb35035