[core] Do not copy stream if all bindings are constructor args
When deserializing an entity which is instantiated via a constructor
with parameters in the general case we need to make a copy of the
input stream so that the constructor arguments can eb deserialized
first. However, when all the serialized values are constructor arguments
we do not need to make a copy as it is slow and inefficient.

closes https://github.com/JetBrains/intellij-community/pull/2232

GitOrigin-RevId: 206b09d385aef797ab0c06b40d177ee91878ff40