Do not copy stream if all bindings are constructor args
When deserializing an entity which is instantiated via a constructor
with parameters in the general case we need to make a copy of the
input stream so that the constructor arguments can eb deserialized
first. However, when all the serialized values are constructor arguments
we do not need to make a copy as it is slow and inefficient.