IDEA-306137 Split start run history and run process status service
The original run configuration is storing now in the recent run history. It solves the problem with wrong recent run history after compound configuration started. Also refactored and simplify the code and listeners logic.

Here may be some regression in the old run widget for the new UI. But now the new one is used.

IJ-CR-98586

GitOrigin-RevId: 54f53aaa0f6ae60dc25af97e35a934f687d97a43