[new UI] IDEA-303508 Fix tool window header button visibility changes
The reason for the buttons (Hide, Options etc.) not always showing on mouse
hover was how their animation was handled. There were three cases:

- If they're not fully visible, but need to be visible, run the showing animation.
- If they're not fully hidden, but need to be hidden, run the hiding animation.
- Otherwise, just cancel all pending animations.

However, animations were synchronized with the actual isVisible property
of the corresponding Swing component, and that synchronization was flawed:

- In the first case, the component is made visible as soon as the animation starts.
- In the second case, the component is made invisible as soon as the animation
EITHER finishes OR is cancelled.
- In the third case, visibility is not updated at all. It's just assumed to be correct.

In most cases it worked just fine, as the last case happens is when
visibility already matches the target value. But if there are pending show/hide
animations, then there was a race condition. Consider a case when a hide animation
was scheduled, and then visibility is immediately set to true
(the mouse leaving and then immediately re-entering the window):

- If the hide animation is already started, it would be replaced by the show animation,
the component visibility would be set to true as soon as it starts, and it
would complete almost immediately (as the button have only barely started to
fade out). The good case.
- If the hide animation is not even started yet, then it would be cancelled, the visibility
would be updated as a result of animation being cancelled, and therefore will be set to false.

So we had an inconsistency here: if the button is hidden or shown using an animation,
however short it may be, then visibility would be updated, but if it's a visual no-op,
then isVisible would be stuck to the previous value.

Fix by:
- checking whether the visibility value associated with the animation
(false for hide, true for show) is still valid when the animation callback is invoked
(completed or cancelled in case of hide, started in case of show);
- and updating visibility in the third case, when the animations are just cancelled.

Now in the third case we are guaranteed to have visibility set correctly and
whatever later animation callback is invoked is guaranteed not to set it to an
incorrect value, assuming it doesn't change again.

GitOrigin-RevId: 8accb10dc4f6d77e9f27dd12884b202ddc5ddf74