[workspace file index] optimize performance and memory usage
In most of the cases, there is a single StoredFileSet for each file in WorkspaceFileIndexData::fileSets, so if we don't wrap them into lists, we'll save a considerable amount of memory in big projects. This change introduces StoredFileSetCollection, which is used to store one, two or several file sets associated with a file. Also, WorkspaceFileIndexData.getFileInfo is optimized to avoid allocating new collections and creating lambdas at least in the most common cases.

For intellij project this reduces the size of memory retained by WorkspaceFileIndexData::fileSets map from 1.3Mb to 600Kb. The performance of ProjectFileIndex's methods improved by around 10%.

GitOrigin-RevId: a388b4cd29bdc02c7df8e025ed76a44444be89a4