[platform] IDEA-305055 Remove all potentially slow code from TreeState.calcId()
Time for some Git/YouTrack archeology!

It all started in 2005 with IDEA-29734 (IDEADEV-2150). A very edge case:
two folders in the project tree having the same name and the same parent
because they are added directly to the project as source roots. Apparently
back then TreeState wasn't used to just save/restore the state when the
tree is closed/reopened, but more often, which led to something as
ridiculous as expanding two nodes in sync. It was fixed by resolving
the tree node to the underlying object (PsiDirectoryNode -> PsiDirectory)
and using its toString() as an ID (PsiDirectory:<full path>).

As the code evolved it got more and more complicated, and in the end
we lost that semantics at some point, and all that digging through data
structures is now for nothing: even for directories it still resolves to just
the directory's name, which is completely pointless as it's available
right away.

It's obvious now that the original issue of different nodes of the same
type having the same parent should be resolved differently, but for now
we're just ditching all that mess with little to no harm. In fact, it even
improves things in certain cases. For example, when displaying structure
of build.xml files, there are some cases when this potentially slow mess
resolved to just the XML tag name, which is obviously the same for all
nested XML tags of the same type, while simple toString() would resolve
to different strings, such as some unique XML attribute value.

GitOrigin-RevId: a30f5b464198a37524893c4b889a91901e321f72