[platform] IDEA-271491 Make PresentableNodeDescriptor thread-safe
The core cause of all kinds of flickering icons in trees was that PresentableNodeDescriptor
performed a lot of modifications on PresentationData instances visible to EDT, and
these modifications were often (e.g. when using AsyncTreeModel) performed on BGT.
This is, of course, as far from thread-safe as it can possibly get.

The fix includes:

- make both (template and updated) presentations AtomicReferences to ensure safe
publishing (see Java memory model);
- make a copy (using the now fixed clone()) of the presentation before making any
modifications in getUpdatedPresentation() and apply();
- do not publish the updated presentation until it's actually updated;
- as a free bonus: don't clone the 'before' presentation in update() as it's not modified
now.

This, of course, breaks semantics to a certain degree, so if any of those almost-600
PresentableNodeDescriptor subclasses (ab)use object identity (e.g. relying on the
fact that the presentation is updated in-place instead of being replaced with a fresh copy),
then it'll likely break. But such code should be fixed anyway.

Also do note that thread-safety here involves only safe publishing and NOT
safe concurrent modifications. If two threads decide to call update() concurrently,
then it's undefined which one ends up being published, and if those updates involve
apply() calls, these can be freely interleaved with actual updates. But at least
any state visible to EDT at any given moment will be a consistent state, not a
partially updated one.

GitOrigin-RevId: 9fd840af8f344ea0fbb066845f8b53cfaf9fb797