CPP-19409: XDebugger: Persist watches in XDebuggerWatchesManager
Otherwise, for "Attach to Process" debug sessions not having a content
to reuse from the environment, modifications made to watches after the
session ends are not preserved.

If a watch was a cause the debugger crashed, having it reappear after
the user removed it, and crash the session again, can be very annoying.

(cherry picked from commit 5c904915d7bad964341f56305c5a8ac21d122843)

IJ-CR-99030

GitOrigin-RevId: 219f184375a7bd3f8e5e16f2064bf836910e132f