[vcs-log] do not lose errors when creating snapshot visible pack
Snapshot VisiblePack is created to save memory and avoid refreshing an invisible Log tab. However, when DataPack is an instance of ErrorDataPack, it does not have any commits anyway and there is no need to make a snapshot. When a VisiblePack is an instance of ErrorVisiblePack, underlying DataPack still has commits, so a snapshot needs to be created. But in this case, ErrorVisiblePack is required to keep an error message and display it in the table.

IDEA-305863

GitOrigin-RevId: 3c0bc14e6633c134d94fa2a90ce863d134eaa730