RUBY-30596 IDEA-274371 IDEA-289574 Do not pass environment variables with invalid names
According to the bash rules: a word consisting solely of letters, numbers, and underscores, and beginning with a letter or underscore. Names are used as shell variable and function names. Also referred to as an identifier.

IJ-CR-98975

(cherry picked from commit 0741ab54f6bbf3d29991b445321fe1f3d976f731)

GitOrigin-RevId: 875798a73b263f6ad77f318d871ea6c0a68266d3