[Workspace Model] [IDEA-306056] Do not load javaSettingsEntity if it's attributes in iml are empty
This state may happen in case we have an iml in .idea with custom content roots, but the original module was renamed outside of idea. So, we have a kind of broken configuration.
I don't think that this is an ideal solution, probably we should store an information that some content roots are custom and do not try to load iml file in this case as a separate module.

Also: IDEA-306514

GitOrigin-RevId: 220f5a0f962685838e888902cf121d748c5a7e4b