[platform UI] CPP-21042 Fix focus trying to get inside an invisible EditorTextField
When an EditorTextField is invisible, SortingFocusTraversalPolicy still calls
its policy's getDefaultComponent(), which returns the inner editor component.
This completely breaks focus navigation with Tab when there's an invisible
EditorTextField is present in the layout.

Fix by making EditorTextField.Jdk7DelegatingToRootTraversalPolicy check
whether the field is visible and displayable, exactly in the same way as it's
done in java.awt.ContainerOrderFocusTraversalPolicy.

GitOrigin-RevId: 5a5fbf984d127c47e96066dff873ee59568ddcf8