[kotlin] hide the problem of classloading from module classloader on project import
The KotlinJsCompilerNotificationImportListener lives in a Project Model 2.0 module and should be loaded with a classloader of that module.

The lambda class which was passed to the ReadAction.nonBlocking.finishOnUiThread was for some reason loaded with a kotlin plugin classloader which broke the check that it should be loaded from the module classloader. After that, the exception was thrown and project import failed.

The current fix just hides the problem to fix the project import. The original problem should be investigated.

(cherry picked from commit 9aa06b80fb7a55c5ca6ba00478d9e9b2b985053d)

GitOrigin-RevId: 9949a0dfbbe6cfaa4173f2266b555cbffa24036e