Make `SimpleStringPersistentEnumerator` robust against duplicate entries in enumeration file
`SimpleStringPersistentEnumerator#enumerate` says that it accepts `null`s, but in fact during serialization `null` is transformed into `"null"` string. During deserialization we read `"null"` string and it is not equal to `null`. So we add `null`, it is stored as `"null"` and we can have many `"null"` strings in the file.

This commit makes sure that we do not corrupt enumeration file and do not overwrite exising entries when duplicate already present: myForwardState has unique line numbers as key, myInvertedState has (possibly) duplicate value as key. Use myForwardState to calculate next index.

GitOrigin-RevId: c556d33fe4c953701f9be2f51c05a3f230492e20