[kotlin] add capability to skip parameter destructuring calls when stepping into a lambda
When stepping into a lambda with parameter destructuring a user would stop inside the lambda's body, but the parameters would still be uninitialized and couldn't be observed in the debugger. This happened because the debugger stopped on an opcode that comes before `component*` function calls. Now after the 'step into' these `component*` calls will be skipped until all the variables are initialized.

KTIJ-23778 Fixed