[kt, tests] Remove WorkspaceModelPrintingGradleImportingTest
It was used mostly as showcase, and the only non-trivial case it checked
was native distribution commonization (barring the fact that checks were
kinda odd). It will be superseded by modern tests infrastructure. For now,
it is removed, as it was getting into the way of refactorings

(cherry picked from commit a99b14eccfe44f1b0a5f5282f93e87fad1ec264a)

GitOrigin-RevId: 511a5fc142bd0f17006510499a90c6ef77d10679