[workspace model] API: add special property to get current snapshot of the workspace storage
It appears that almost all usages of `WorkspaceModel::entityStorage` property just take `current` property from the result. Adding a special property for that simplifies code, and mentioning 'snapshot' word in the property name and in its type emphasizes the fact that the returned value is a snapshot which won't be affected by future changes.

GitOrigin-RevId: 6ec75a6ad826f864256c04b27dfe3ff72f51c297