Fix MessageBus listener leak in WelcomeFrame
The current code assumes that the frame is only disposed by receiving a
projectOpened event. However, it is possible to call dispose() directly;
this is done by Android Studio, which uses two WelcomeFrames sequentially
on a fresh install.

When dispose() is called directly, the MessageBus listener is not
removed; this is a memory leak. It also causes a NullPointerException
when the projectOpened event occurs later: the dangling listener calls
resetInstance(), which causes FlatWelcomeFrame's cleanup method to fail.