Fix memory leak in WelcomeFrame when dispose method is called directly
The current code assumes that the frame is only disposed by receiving a projectOpened event. However, it is possible to call dispose() directly. This is done by Android Studio, which uses two WelcomeFrames sequentially on a fresh install.

When dispose() is called directly, the MessageBus listener is not removed. This is a memory leak. It also causes a NullPointerException when the projectOpened event occurs later: the dangling listener calls resetInstance(), which causes FlatWelcomeFrame's cleanup method to fail.

closes https://github.com/JetBrains/intellij-community/pull/2259

(cherry picked from commit 9654c6bc0ff1bdb93e1ac64f5e925c807de4fadb)

Change-Id: I9654c6bc0ff1bdb93e1ac64f5e925c807de4fadb