Fix memory leak in WelcomeFrame when dispose method is called directly
The current code assumes that the frame is only disposed by receiving a projectOpened event. However, it is possible to call dispose() directly. This is done by Android Studio, which uses two WelcomeFrames sequentially on a fresh install.

When dispose() is called directly, the MessageBus listener is not removed. This is a memory leak. It also causes a NullPointerException when the projectOpened event occurs later: the dangling listener calls resetInstance(), which causes FlatWelcomeFrame's cleanup method to fail.

closes https://github.com/JetBrains/intellij-community/pull/2259

GitOrigin-RevId: f2896f4b0123ee82898aa31850a223e777d9492c