Add EP for override member annotations
This fix addresses https://youtrack.jetbrains.com/issue/KTIJ-21347.

The "Override Member" action currently removes all annotations except
for those marked with @RequireOptIn. This change adds an extension point
that allows plugins to specify whether any other annotations should be
retained as well.

Since there is already a java `OverrideImplementsAnnotationsHandler`,
this change refactors out the common applicable portion of that handler
to a super interface and extension point in the platform that can be
reused by the Kotlin plugin.

Note that the K2 version of Override Members would not show this
behavior, since it actually retains all annotations. This is likely an
unintentional change, and I filed
https://youtrack.jetbrains.com/issue/KTIJ-23517 to track that.