KTIJ-23683: Sort descriptors for single imports
https://youtrack.jetbrains.com/issue/KTIJ-23683

SingleImportVariant effectively displays the first descriptor in the
list it's given (with the caveat that it prefers ClassDescriptors
over other descriptors). This changes sorts the list before it's given
to SingleImportVariant, using the priority that's already being
calculated for the descriptors. Doing this ensures the highest priority
(or rather, lowest numeric value priority) is used for display.

closes https://github.com/JetBrains/intellij-community/pull/2235

(cherry picked from commit c5b047368c3a23917771aa63d629e74319fa9acc)

Change-Id: Iafcad7e9efcfa588a54e485055de6b4446f06520