KTIJ-23683: Sort descriptors for single imports
https://youtrack.jetbrains.com/issue/KTIJ-23683

SingleImportVariant effectively displays the first descriptor in the
list it's given (with the caveat that it prefers ClassDescriptors
over other descriptors). This changes sorts the list before it's given
to SingleImportVariant, using the priority that's already being
calculated for the descriptors. Doing this ensures the highest priority
(or rather, lowest numeric value priority) is used for display.

closes https://github.com/JetBrains/intellij-community/pull/2235

(cherry picked from commit 5020f73615b84954810b414a2c69254f5ac57520)

GitOrigin-RevId: 56d4bf98c4cb3a5bfffdf4833cd38c63665ce547