[UI] IDEA-298485 Create GeometryService
To implement Compact Mode, gather all geometry-related
parameters in one place, so we can easily change them
depending on what mode is in effect.

Use UI Defaults as a backing storage, as this way we can
test changes interactively by using the LaF Defaults internal
action.

Note that UIManager.getDefaults() is used instead of
getLookAndFeelDefaults(). This is because it's used
by the LaF Defaults internal action as well (unlike
LafManagerImpl.updateUI). It isn't really important
for GeometryService.

As a demonstration, change
ActionToolbar.EXPERIMENTAL_TOOLBAR_MINIMUM_BUTTON_SIZE
to use the new API. It isn't applied dynamically yet, as this value
is only used in the constructor, but it already can be checked
by reopening the project after editing LaF Defaults.

GitOrigin-RevId: fc67be8e4b2b4270e411a8c937658bb9f1bb600c