[workspace model] Fix restoring references to new parents for unchanged entities
Before the fix we can't truly detect new parents if they were not handled. Thus, such entities from replace storage were not marked as `AddedElement`.  We make an assumption that if the parent of parent was not changed, and we can't find parent of child in target storage thus it's a new parent for child.

GitOrigin-RevId: 62cc81698e521054a4eb563f58ff23765860f1b9