IDEA-307889 Fix absence of 'show hidden tabs' button when 'Show pinned tabs in a separate row' enabled
There was the problem when table layout tried to calculate more toolbar width, but it returned 0, because TabListAction contained in the toolbar was invisible in the result of the update that rely on JBTabsImpl#canShowMorePopup, that checks that more toolbar rect from the previous layout is not empty.
So, there was the sort of cycle: toolbar pref size is zero, because the contained action is not visible, but action is not visible because toolbar size is zero.
So, I decided to hardcode more toolbar width. It should be safe, because there is always only one button inside.

(cherry picked from commit a927408e89ea38a9bb8b76c92a79890cd480aefc)

IJ-CR-100112

GitOrigin-RevId: 181b4c5c894ee1706fe0c184a42a547228ee9820