settings-sync: during regular sync, push local settings if there is no file on server
If the file is not deleted intentionally (which intentional deletion is actually indicated by an empty snapshot with a special marker) and just doesn't appear on the server for some reason, then the regular sync settings procedure should push the file back to the server.

An example of such situation is disabling the cross-ide-sync (IDEA-307136), which changes the target location of the file on server, and that location is initially empty.

(cherry picked from commit 9eeba757ba10047e9a7c47f200923090ed65ba2d)

IJ-CR-100596

GitOrigin-RevId: c33b67c5b0d25cbec062c646e85f452fd07d4879