[UI] IDEA-307246 Store unified weights in window layouts
If a saved layout doesn't have any tool windows docked
to a certain anchor, then that anchor's area weight would
be reset on the Restore Layout action. This is most likely
undesirable behavior. Consider the following: the user
opens the project tool window, adjusts its width, hides it
and stores the layout. When the layout is restored, it is
expected that activating the project tool window would
have the same size it had before saving the layout,
even though it wasn't visible when the layout was stored.

Fix by storing unified widths (weights) of the docking
areas alongside with tool window layouts.

GitOrigin-RevId: b9e143009185b1758bc0e98cf48ce29f3098105d