[java-analysis] Prefer nullity from contract over nullity from default type annotation
Under @DefaultQualifier, all simple type instances are assumed to be notnull. Type of argument for Object obj = Objects.requireNonNull(xyz) is inferred to be `Object` type mention. As a result, it's believed to be notnull.

While we are not fixing the issue in general, in this particular case, we can disable type-based nullability, as we know requireNonNull contract

Fixes IDEA-303382 Incorrect inspection for requireNonNull with DefaultQualifier annotation

GitOrigin-RevId: fa010ac638be4a78b9464c1a6faa4d4d7f08ec34