PY-57820 Obtain proper language level for Python 2 interpreters
Python interpreter version output is parsed in `PythonSdkFlavor.getLanguageLevelFromVersionString()`. This method uses an abstract `PythonSdkFlavor.getName()` method to construct the pattern for parsing. Prior to these changes `UnknownFlavor` was used by default for the detected Python interpreters found via `detectSystemWideSdks()` method in PySdkExt.kt. The implementation of `getName()` method in `UnknownFlavor` returns an empty string. This causes inability to match the Python version output in `PythonSdkFlavor.getLanguageLevelFromVersionString()` method. To overcome this problem the actual `PythonSdkFlavor` is now stored in `PyDetectedSdkAdditionalData`, which allows to match the version output correctly.

GitOrigin-RevId: cbe6fffaaf90254d195afac0ec59ff34dd319247