settings-sync: make cross-ide sync optional and disable it by default: IDEA-233535
If cross-ide sync is disabled (which is by default since this commit), then the settings-snapshot.zip is stored on the server under a subfolder correspondent to the current IDE (e.g. '/idea/'). Otherwise, the snapshot is stored in the root, as it were previously.

The "cross-ide" property is stored in a marker file '/cross-ide-sync-enabled' located at the server root. File exists => cross-ide sync is enabled, file doesn't exist => cross-ide sync is disabled. Locally the flag is stored in SettingsSyncLocalSettings since there is not much sense to sync it: if modified, it changes the location of the snapshot, and other IDEs can't know about it before knowing the state of the flag.

Before each sync procedure, the IDE checks the state of the file. If it is different from what this IDE knows, the local flag is updated, and the new location is used to sync.

When the user changes the flag from the Settings, we not only update the marker file on the server, but also force push current snapshot to the new place to overwrite any potential old snapshot residing there. To do it we handle the CrossIdeSyncStateChanged separately from other events, not to interfere with the regular sync procedure.

GitOrigin-RevId: 2b388444c3eec8a7d867c5e969bf2e3373003476