Delete mini-batches in `UnindexedFilesIndexer.indexFiles` to increase CPU load
There is no need to organize files for indexing in mini-batches and wait for each batch completion. All files for indexing are now in a single queue. Progress indicator works similar to profiler in sampling mode: more tasks we have from the same provider, the higher probability that corresponding text will be visible in UI progress.

GitOrigin-RevId: 2f2269d1f01879d366e28b8a9099d50a6406d5a5