[UI] IDEA-304712 Move appendInplaceComments to BGT update
The appendInplaceComments method used to be called on the EDT
during painting. It's potentially slow, and it's also a bad practice
to do any state computations during painting anyway.

Fix by moving the relevant logic into the node's update() and
appending the comments directly to PresentationData.

This turned out to be trickier than it seems. The Project view has a lot
of different node types, some of which are even in 3rd party plugins.
They all extend ProjectViewNode (except Rider that extends AbstractTreeNode).
However, it's not easy to put the logic there as it's a part of the lang-api module,
while the logic itself is in the lang-impl module. And logic doesn't belong to the API
module anyway.

The solution is, of course, dependency inversion, but it's not easy to inject a reference
to an interface into ProjectViewNode (let alone AbstractTreeNode) because
we have no control of how these objects are constructed.
Of course, it's possible to delegate this job to some parent or owner
object like the Project View Pane, for example. However, this creates unnecessary coupling
and would require some dirty hacks to get to those nodes and inject the needed references
into them.

Thankfully, most (if not all) Project View nodes that actually use inplace comments
extend AbstractPsiBasedNode, so we just put the actual comment producer there.

The new interfaces for this DI is InplaceCommentProducer (that produces the actual
comments) and InplaceCommentAppender that appends them (either to the presentation
when the new approach is used or to the renderer for the legacy case).

In case there are nodes that don't provide an inplace comment producer,
we still revert to the old logic. This has an unfortunate side effect of calling
getValue() on a node, which is the very potentially slow operation we're trying
to get rid of, but at least we now don't call it on AbstractPsiBasedNode anymore,
and those calls were probably the vast majority (if not all) slow operation calls.

One more trick involved is to convert the plain presentation's text (myPresentableText)
to the colored text before appending comments because many presentations don't
use colored text by default. Just appending comments would override the plain text
and, e.g., class and file names in the Project View would disappear. This is what the new
ensureColoredTextIsUsed() method in PresentationData is for.

GitOrigin-RevId: d7da0fc059d6d70ff369ea981501c5008b53b20d