IDEA-272466 Fix showing close button for non-closeable tabs
If we obtain data context from the button, it will return data from JBTabsImpl. But context of specific tab is required here. The problem of this approach is that actions won't be performed when tab content component is not showing, but it is solved by adding ALLOW_ACTION_PERFORM_WHEN_HIDDEN client property to the tab content component.

GitOrigin-RevId: 18a9f2f298375a316e9147d44e629ed178cfce6f