code stlye: remove maximum right margin
This option is really annoying, I sometimes have lines that exceed that value (e.g. minified files).
I never ever want them to be hard wrapped, this option allows this.
If you do not need it, simply set the maximum hard wrap not that high, leave it at the default value (120)