IDEA-304614 don't register listener onto a disposed lookup, cancel the popup request instead
The listener was never invoked causing the leak.
(The lookup should've thrown an assertion on adding a listener if the lookup is already disposed.)

Also, after this change, documentation browser/popup are disposed if cancellation happens before ownership transfer:
- Browser is disposed if cancellation happens before it's attached as a child disposable to `DocumentationPopupUI`,
which is attached as a child disposable to the popup.
- Popup is disposed if cancellation happens before it's returned to the caller.

(cherry picked from commit 9a0440f06dd80201f504f57eb29766cd20088e64)

Bug: 252870216
Change-Id: I9a0440f06dd80201f504f57eb29766cd20088e64