CIDR: RemoteUnixPipe: Fix ProcessWaitFor leak (missing startNotify() call)
Calling startNotify() on a BaseOSProcessHandler is crucial because otherwise
the ProcessWaitFor thread launched to wait until the process terminates
would block indefinitely in an attempt to get a termination callback,
which is supposed to be installed from startNotify().

This change aims to fix a project leak reproducing on WSL1 configuration,
although this thread leak in question might only be a part of the problem.

GitOrigin-RevId: 359a52b157ad08f84fb5d51f15392ae180ead2f2