settings-sync: disable sync of plugins for users who didn't use the new settings sync yet: IDEA-308614
We aren't changing the default value of the Plugins checkbox for two reasons:
1. Not to hurt existing happy users of 2022.3.0 which sync plugins (changing default would disable plugin sync for them).
2. We're planning to change the 'default' back to enabled after we implement cross-ide sync of plugins.

Instead, we disable the checkbox in all cases when the settings sync can be enabled manually or automatically for the first time (to preserve the user setting if they do enable-disable-enable).

There is existing flag to understand the "first time user" - migrationFromOldStorageChecked allows to execute the code only once.

GitOrigin-RevId: 2943be75be3878aa9b95609e785c491228dd6d29