settings-sync: update categories checkboxes from their real data in the settings: IDEA-308723
SettingsDescriptors, used as a model for checkboxes, are initialized always with 'true'.
onReset() updates SettingsDescriptors by the data from settings.
SettingsDescriptors are bind to the checkboxes, so onReset() the checkboxes are updated by the state of SettingsDescriptors.

But the order of these onReset() is undefined, and in fact firstly checkboxes are updated by the always-false state of SettingsDescriptors, and only then the descriptors themselves get updated from the settings.

The fix is to update all descriptors by actual values from settings before building the panel.

GitOrigin-RevId: f95951fc543f412f08c79fdac3c05e990b67f12d