settings-sync: check if sync is needed, in the main settings sync thread
Previously it was done in a separate thread, which complicated the flow and led to various issues (including IDEA-308972, IDEA-302107).

Now, when the time comes to make a sync check (e.g. by timer of on frame activation), we send a request which is processed in the same queue as all other requests, and, most importantly, after the initialization.

SyncRequest is different from other requests, which modify the settings sync history. We don't want it to be processed in the middle of processing apply-from-server and other similar requests. Instead, we process it in the same queue, but separately, just like the cross-ide-sync check is processed, as so called 'exclusive' events.

GitOrigin-RevId: 760a738e56e499c4ba16ca04ac75bf5e03612e5e