[workspace model] store entities corresponding to unloaded modules in a separate storage (IDEA-298694)
Most of the code should ignore entities corresponding to unloaded modules, don't react to changes in them, if an entity becomes unloaded,
it usually should be processed the same way as it is removed, and if an entity is loaded back, it usually should be processed the same way
as it is added. This is now achieved by storing unloaded entities in a separate versioned storage in WorkspaceModel. When a module is unloaded,
the corresponding entities are removed from the main storage and added to the separate one. When a module is loaded back, its entities
are removed from the separate storage and added to the main storage. There is now a special topic (WorkspaceModelTopics.UNLOADED_ENTITIES_CHANGED)
for those who are interested in changes in unloaded entities.
Code which loads project configuration from .idea, puts loaded and unloaded entities to separate builders. Code which saves project configuration,
takes the entities from both storages. Unloaded entities are stored in a separate cache file.

(Cherry-picked from 6039f341ef0a44b871ebb514ab9cf29d060ce2ba, IJ-CR-100740)

GitOrigin-RevId: fbc36d1390f4d3030360511ddeb64f53f98e606f